chore: better defaults for editor marks
- inclusive: means cursor placed at the end of mark will still type
inside the mark. This is kind of not good for smartphones, but we
provide easy way to disable it where you need with inline formatter.
- exitable: this option is irrelevant for phone users, but for desktops
it enables to move out of mark with keyboard arrows. Which i think is
the best solution because in large screens we showing formatter only at
selection.